Fix Ctrl-C not abandoning a pomodoro sometimes
It's possible for Ctrl-C to not have the intended effect of abandoning
the current pomodoro.  This is subject to pretty specific timing and
chance: the interrupt has to be received just as the time is changed.
The bug is caused by any function that was invoked by the `pomodoro()`
function being executed when SIGINT is received.  The inner function
returns 1 (SIGINT trap), but the `pomodoro()` function does not return;
the `state` variable is set to 0, but the timer continues.

When Ctrl-C is pressed again after this happened, the program terminates
rather than only abandoning the current pomodoro and prompting the user
to restart it.

This commit hopefully fixes all of that.